EES 2.0+ / Added window and openGL stuff. Added a class for writing text in the window.

+++ src/elevatordata.cpp
@@ -189,7 +189,7 @@ int ElevatorData::dystans(int w, char direction)
         distance=abs(last_floor_number-polecenia[0]);
     }
 
-    for(int j=0;j<(polecenia.size()-1);j++)
+    for(size_t j=0;j<(polecenia.size()-1);j++)
     {
         if((polecenia[j]<w && polecenia[j+1]>w && direction=='u') || (polecenia[j]>w && polecenia[j+1]<w && direction=='d'))
         {
